REF: move remaining offsets into liboffsets (#34398)
