+++ doc/source/reference/offset_frequency.rst
@@ -134,6 +134,7 @@ Methods
 .. autosummary::
    :toctree: api/
 
+    CustomBusinessDay.apply_index
     CustomBusinessDay.apply
     CustomBusinessDay.copy
     CustomBusinessDay.isAnchored
