+++ pandas/tests/indexes/datetimes/test_date_range.py
@@ -10,6 +10,7 @@ import pytz
 from pytz import timezone
 
 from pandas._libs.tslibs import timezones
+from pandas._libs.tslibs.offsets import BDay, CDay, DateOffset, MonthEnd, prefix_mapping
 from pandas.errors import OutOfBoundsDatetime
 import pandas.util._test_decorators as td
 
@@ -17,8 +18,6 @@ import pandas as pd
 from pandas import DatetimeIndex, Timestamp, bdate_range, date_range, offsets
 import pandas._testing as tm
 from pandas.core.arrays.datetimes import generate_range
-
-from pandas.tseries.offsets import BDay, CDay, DateOffset, MonthEnd, prefix_mapping
 
 START, END = datetime(2009, 1, 1), datetime(2010, 1, 1)
